Extract SearchPane to separate file
