+++ src/App.js
@@ -1,10 +1,9 @@
 import React from "react";
 import { useState, useEffect } from "react";
 import config from "./pages.json";
+import { SearchPane } from "./SearchPane.js";
 import styled from "styled-components";
-import { ChevronLeft, ChevronRight, Rewind, Search } from "@core-ds/icons/16";
-import Modal from "@material-ui/core/Modal";
-import Paper from "@material-ui/core/Paper";
+import { ChevronLeft, ChevronRight, Rewind } from "@core-ds/icons/16";
 import LinearProgress from "@material-ui/core/LinearProgress";
 import Button from "@material-ui/core/Button";
 import TextField from "@material-ui/core/TextField";
@@ -66,72 +65,6 @@ function Page({ src, width }) {
       </PageLink>
       <PageFrame src={src} width={width} />
     </PageWrapper>
-  );
-}
-
-const ResultLink = styled.a`
-  display: block;
-  padding-bottom: 4px;
-`;
-
-const Scroller = styled.div`
-  width: 100%;
-  flex-grow: 1;
-  overflow: scroll;
-`;
-
-function Results({ search, close }) {
-  const re = new RegExp(search, "i");
-  return (
-    <Scroller>
-      {pages.map((page, idx) => {
-        if (re.test(page)) {
-          return (
-            <ResultLink onClick={close} href={`#${idx}`}>
-              {page}
-            </ResultLink>
-          );
-        }
-      })}
-    </Scroller>
-  );
-}
-
-const SearchDialog = styled(Paper)`
-  position: absolute;
-  width: max-content;
-  max-width: 80%;
-  max-height: 80%;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 16px;
-`;
-
-function SearchPane() {
-  const [open, setOpen] = useState(false);
-  const [search, setSearch] = useState("");
-  const updateSearch = (evt) => {
-    setSearch(evt.target.value);
-  };
-  const closeModal = () => setOpen(false);
-  return (
-    <>
-      <Button onClick={() => setOpen(true)}>
-        <Search />
-      </Button>
-      <Modal open={open} onClose={closeModal}>
-        <SearchDialog>
-          <div>
-            <input type="text" autoFocus onChange={updateSearch}></input>
-          </div>
-          <Results close={closeModal} search={search} />
-        </SearchDialog>
-      </Modal>
-    </>
   );
 }
 
@@ -224,7 +157,7 @@ function App() {
             onChange={updateWidth}
             value={width}
           />
-          <SearchPane />
+          <SearchPane pages={pages} />
         </Controls>
       </Header>
       <Comparison>
